requirements.txt
# Van Gogh Living Scene — Python dependencies
# Install inside venv with --system-site-packages (for picamera2)
#
# Python 3.13+ on aarch64 (Raspberry Pi OS Bookworm)
# Last verified: 2026-04-01
#
# SEC-04: To generate hash-pinned requirements on the target:
#   pip install pip-tools
#   pip-compile --generate-hashes requirements.in -o requirements.lock
#   Then update install.sh to use: pip install --require-hashes -r requirements.lock
#
# SEC-05: Pillow floor bumped to >=12.0 for Python 3.13 wheels + CVE fixes (OWASP A06)
#
# Migration notes:
#   tflite-runtime replaced by ai-edge-litert (Google's successor, drop-in API)
#     Import changes: tflite_runtime.interpreter -> ai_edge_litert.interpreter
#   numpy upgraded to 2.x (required for Python 3.13 wheels)
#   inky[rpi] extra removed (no longer exists in inky>=2.0)
#
# References (as of 2026-04-01):
#   rembg 2.0.74       — https://pypi.org/project/rembg/           — 2026-03-25
#   ai-edge-litert 2.1.3 — https://pypi.org/project/ai-edge-litert/ — 2026-03-16
#   Pillow 12.2.0       — https://pypi.org/project/Pillow/          — 2026-04-01
#   numpy 2.4.4         — https://pypi.org/project/numpy/           — 2026-03-29
#   PyYAML 6.0.3        — https://pypi.org/project/PyYAML/          — 2025-09-25
#   inky 2.3.0          — https://pypi.org/project/inky/            — 2026-01-29

rembg[cpu]>=2.0.74,<3.0
ai-edge-litert>=2.1.3
Pillow>=12.0,<13.0
inky>=2.0,<3.0
PyYAML>=6.0,<7.0
numpy>=2.2,<3.0
